[Fix] #PO-26 - DataModel Hashable하게 변경
- Device는 각각 고유한 UUID로 Hashable하게 받아서 DataSource diffing 머신을 활용해서 실시간으로 UI를 바꾸려고 변경
